requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.1.7
django-multiselectfield==0.1.8
django-phone-field==1.7.1
djangorestframework==3.9.3
httpie==1.0.2
idna==2.8
Pillow==6.0.0
psycopg2==2.7.4
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.3
